+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 - **[Feature]** Provide Delayed Topics (#1000)
 - **[Feature]** Provide ability to expire messages (expiring topics)
 - **[Feature]** Provide ability to apply filters after messages are polled and before enqueued. This is a generic filter API for any usage.
+- [Improvement] When using ActiveJob with Virtual Partitions, Karafka will stop if collectively VPs are failing. This minimizes number of jobs that will be collectively re-processed.
+- [Improvement] `#retrying?` method has been added to consumers to provide ability to check, that we're reprocessing data after a failure. This is useful for branching out processing based on errors.
 - [Improvement] Track active_job_id in instrumentation (#1372)
 - [Improvement] Introduce new housekeeping job type called `Idle` for non-consumption execution flows.
 - [Improvement] Change how a manual offset management works with Long-Running Jobs. Use the last message offset to move forward instead of relying on the last message marked as consumed for a scenario where no message is marked.